pyproject.toml
[tool.poetry]
name = "tutor-platform"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tutor_platform"}]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.3"
bcrypt = "^4.0.1"
aiopg = "^1.4.0"
uuid = "^1.30"
redis = "^4.4.2"
pyjwt = "^2.6.0"
python-dotenv = "^0.21.0"
asyncio = "^3.4.3"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.6.5"
ipykernel = "^6.20.1"
notebook = "^6.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
